StringX done!
There is actually an error in the problem description. You don't actually care if the first and last char is x. For example, what happen if my string is xabcd this will return the same string. So, it does not really matter you are only checking if the next to the first char to the char before the last char contains "x" if that is the case removed it using ! in front of your if.